// frontend-admin/src/api/generated/model/userInfoDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */

export interface UserInfoDto {
  /** @nullable */
  id?: string | null;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  role?: string | null;
}